\name{PELVIS-package}
\alias{PELVIS-package}
\alias{PELVIS}
\docType{package}
\title{
Probabilistic Sex Estimate using Logistic Regression, Based on VISual Traits of the Human Os Coxae}
\description{
An R-Shiny application implementing two methods for sexing the human os coxae, based on eleven trichotomic traits.
}
\details{
\tabular{ll}{
Package: \tab PELVIS\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2019-03-11\cr
License: \tab GPL >=2\cr
}

}
\author{
Frédéric Santos, <frederic.santos@u-bordeaux.fr>
}
\references{
Bruzek, J. (2002) A method for visual determination of sex, using the human hip bone. \emph{American Journal of Physical Anthropology} \bold{117}, 157--168. doi: 10.1002/ajpa.10012

Santos, F., Guyomarc'h, P., Rmoutilova, R. and Bruzek, J. (Submitted to \emph{American Journal of Physical Anthropology}) A method of sexing the human os coxae based on logistic regressions and Bruzek's nonmetric traits
}
\keyword{bruzek, hip bone, logistic regression, morphoscopy, os coxae, sex estimate, sex estimation}
\examples{
if(interactive()){ StartPELVIS() }
}
