\name{PELVIS-package}
\alias{PELVIS-package}
\alias{PELVIS}
\docType{package}
\title{
  Probabilistic Sex Estimate using Logistic Regression, Based on VISual
  Traits of the Human Os Coxae
}
\description{
  An R-Shiny application implementing a method of sexing the human os
  coxae based on logistic regressions and Bruzek's nonmetric traits.
}
\details{
  \tabular{ll}{
    Package: \tab PELVIS\cr
    Type: \tab Package\cr
    Version: \tab 2.0.0\cr
    Date: \tab 2019-07-22\cr
    License: \tab CeCILL-2.1\cr
  }
}
\author{
  Frédéric Santos, <frederic.santos@u-bordeaux.fr>
}
\references{
  Bruzek, J. (2002) A method for visual determination of sex, using the
  human hip bone. \emph{American Journal of Physical Anthropology}
  \bold{117}, 157--168. doi: 10.1002/ajpa.10012

  Santos, F., Guyomarc'h, P., Rmoutilova, R. and Bruzek, J. (2019) A
  method of sexing the human os coxae based on logistic regressions and
  Bruzek's nonmetric traits. \emph{American Journal of Physical
  Anthropology} \bold{169}(3), 435-447. doi: 10.1002/ajpa.23855
}
\examples{
if(interactive()){ StartPELVIS() }
}
\concept{hip bone}
\concept{os coxae}
\concept{ossa coxae}
\concept{sex estimation}
\concept{biological anthropology}
\concept{morphoscopy}
\concept{bruzek}
